move Scorpio.stringify_symbol_keys, Scorpio::FingerprintHash, Scorpio::Memoize to scorpio/util.rb
